use base activities for all types of routers
