+++ examples/src/main/java/com/mapbox/navigation/examples/core/HybridRouterActivityJava.java
@@ -1,0 +1,10 @@
+package com.mapbox.navigation.examples.core;
+
+import com.mapbox.navigation.base.route.Router;
+
+public class HybridRouterActivityJava extends BaseRouterActivityJava {
+
+  @Override Router setupRouter() {
+    return setupHybridRouter(this);
+  }
+}
