+++ examples/src/main/java/com/mapbox/navigation/examples/core/HybridRouterActivityKt.kt
@@ -1,0 +1,7 @@
+package com.mapbox.navigation.examples.core
+
+import com.mapbox.navigation.base.route.Router
+
+class HybridRouterActivityKt : BaseRouterActivityKt() {
+    override fun setupRouter(): Router = setupHybridRouter(this)
+}
