+++ examples/src/main/res/values/strings.xml
@@ -20,6 +20,12 @@
     <string name="title_onboard_router_java" translatable="false">Onboard Router Java</string>
     <string name="description_onboard_router_java" translatable="false">Shows routes using onboard routing api in Java code.</string>
 
+    <string name="title_hybrid_router_kotlin" translatable="false">Hybrid Router Kotlin</string>
+    <string name="description_hybrid_router_kotlin" translatable="false">Shows routes using hybrid routing api in Kotlin code.</string>
+
+    <string name="title_hybrid_router_java" translatable="false">Hybrid Router Java</string>
+    <string name="description_hybrid_router_java" translatable="false">Shows routes using hybrid routing api in Java code.</string>
+
     <string name="title_trip_service_kotlin" translatable="false">Trip Service Kotlin</string>
     <string name="description_trip_service_kotlin" translatable="false">Shows how to use the Service to push notifications in Kotlin code.</string>
